+++ src/components/CheckboxList.js
@@ -1,4 +1,3 @@
-import React, { useState, useEffect } from "react";
 import { Typography } from "@mui/material";
 import FormControlLabel from "@mui/material/FormControlLabel";
 import Checkbox from "@mui/material/Checkbox";
@@ -6,14 +5,35 @@ import DeleteIcon from "@mui/icons-material/Delete";
 import IconButton from "@mui/material/IconButton";
 import Paper from "@mui/material/Paper";
 import Grid from "@mui/material/Grid";
+import LoadingSpinner from "./spinner";
 
-export default function CheckboxList({ Tasks, deleteTask }) {
+export default function CheckboxList({
+  Tasks,
+  deleteTask,
+  isLoading,
+  isError,
+}) {
   const delTask = (id) => deleteTask(id);
+
+  if (isLoading) {
+    return (
+      <div>
+        <LoadingSpinner />
+      </div>
+    );
+  }
+  if (isError) {
+    return (
+      <div>
+        <h1>Error...</h1>
+      </div>
+    );
+  }
 
   const checkTask = (id) => {
     Tasks.map((task) => {
       if (task.id === id) {
-        task.iscompleted = !task.iscompleted;
+        task.completed = !task.completed;
       }
     });
   };
@@ -21,7 +41,7 @@ export default function CheckboxList({ Tasks, deleteTask }) {
   return (
     <>
       {Tasks.map((task) => {
-        const { id, title, iscompleted } = task;
+        const { id, title, completed } = task;
         return (
           <Paper variant="outlined" key={id}>
             <Grid container>
@@ -33,7 +53,7 @@ export default function CheckboxList({ Tasks, deleteTask }) {
               <Grid item xs={8}>
                 <Typography
                   style={
-                    iscompleted
+                    completed
                       ? { textdecoration: "line-through" }
                       : { textdecoration: "none" }
                   }>
