+++ src/spinner.css
@@ -10,18 +10,17 @@
 
 .loading-spinner {
     display: inline-block;
-    width: 100px;
-    height: 100px;
+    width: 80px;
+    height: 80px;
     border: 10px solid #f3f3f3;
     /* Light grey */
     border-top: 10px solid #383636;
     /* Black */
     border-radius: 50%;
     animation: spinner 1.5s linear infinite;
-    position: absolute;
-    left: 45%;
+    position: relative;
     margin: 0px auto;
-    margin-top: 20%;
+    margin-top: 5%;
 }
 
 .spinner-container {
